Websocket handshake refactoring.
Addresses https://github.com/dotnet/aspnetcore/issues/31373
