+++ src/Middleware/WebSockets/src/HandshakeHelpers.cs
@@ -2,7 +2,6 @@
 // Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
 
 using System;
-using System.Collections.Generic;
 using System.Security.Cryptography;
 using System.Text;
 using Microsoft.AspNetCore.Http;
@@ -34,61 +33,6 @@ namespace Microsoft.AspNetCore.WebSockets
         };
 
         // Verify Method, Upgrade, Connection, version,  key, etc..
-        public static bool CheckSupportedWebSocketRequest(string method, List<KeyValuePair<string, string>> interestingHeaders, IHeaderDictionary requestHeaders)
-        {
-            bool validUpgrade = false, validConnection = false, validKey = false, validVersion = false;
-
-            if (!string.Equals("GET", method, StringComparison.OrdinalIgnoreCase))
-            {
-                return false;
-            }
-
-            foreach (var pair in interestingHeaders)
-            {
-                if (string.Equals(HeaderNames.Connection, pair.Key, StringComparison.OrdinalIgnoreCase))
-                {
-                    if (string.Equals(HeaderNames.Upgrade, pair.Value, StringComparison.OrdinalIgnoreCase))
-                    {
-                        validConnection = true;
-                    }
-                }
-                else if (string.Equals(HeaderNames.Upgrade, pair.Key, StringComparison.OrdinalIgnoreCase))
-                {
-                    if (string.Equals(Constants.Headers.UpgradeWebSocket, pair.Value, StringComparison.OrdinalIgnoreCase))
-                    {
-                        validUpgrade = true;
-                    }
-                }
-                else if (string.Equals(HeaderNames.SecWebSocketVersion, pair.Key, StringComparison.OrdinalIgnoreCase))
-                {
-                    if (string.Equals(Constants.Headers.SupportedVersion, pair.Value, StringComparison.OrdinalIgnoreCase))
-                    {
-                        validVersion = true;
-                    }
-                }
-                else if (string.Equals(HeaderNames.SecWebSocketKey, pair.Key, StringComparison.OrdinalIgnoreCase))
-                {
-                    validKey = IsRequestKeyValid(pair.Value);
-                }
-            }
-
-            // WebSockets are long lived; so if the header values are valid we switch them out for the interned versions.
-            if (validConnection && requestHeaders[HeaderNames.Connection].Count == 1)
-            {
-                requestHeaders[HeaderNames.Connection] = HeaderNames.Upgrade;
-            }
-            if (validUpgrade && requestHeaders[HeaderNames.Upgrade].Count == 1)
-            {
-                requestHeaders[HeaderNames.Upgrade] = Constants.Headers.UpgradeWebSocket;
-            }
-            if (validVersion && requestHeaders[HeaderNames.SecWebSocketVersion].Count == 1)
-            {
-                requestHeaders[HeaderNames.SecWebSocketVersion] = Constants.Headers.SupportedVersion;
-            }
-
-            return validConnection && validUpgrade && validVersion && validKey;
-        }
-
         public static void GenerateResponseHeaders(string key, string? subProtocol, IHeaderDictionary headers)
         {
             headers[HeaderNames.Connection] = HeaderNames.Upgrade;
